fix failed assertions when running from TTY
When running imv from a TTY, assertions in imv_window_create are hit.
This patch replaces these assertions with propogating the error upward
to setup_window.

It is necessary to change the return value of connect_to_wayland from
void to bool to indicate success, since the operation is fallible.

Signed-off-by: Nissa <and-nissa@protonmail.com>