feat: Wayland cursor support
I was planning to set the grab face when panning the image, but there is
no way to get whether we are actually panning, or using the cursor
motion for something else. While it is true that we only use cursor
motions for panning currently, it feels hacky to rely on that, and might
bite back in the future.